--- subekashi/models.py
@@ -1,163 +1,0 @@
-from django.db import models
-from django.utils import timezone
-
-
-# SongやAuthorを編集したユーザーのIPアドレスの情報
-# 画面上では"全て{song_id}の所為です。"と表示される
-class Editor(models.Model):
-    ip = models.CharField(default = "", unique=True, max_length = 100)
-    is_open = models.BooleanField(default = True)
-    is_forced_open = models.BooleanField(default = False)
-
-    def __str__(self):
-        return f"全て{self.id}の所為です。"
-
-
-# 楽曲のメイン情報
-# 外部キーはまだ無いので後ほどカスタムコマンドで対応を行う
-class Song(models.Model):
-    CHOICES = (
-        ("song", "曲"),
-        ("arrange", "アレンジ"),
-        ("otomad", "音MAD"),
-        ("medley", "メドレー"),
-        ("omnibus", "総集編"),
-        ("notice", "告知"),
-        ("dec", "解読"),
-        ("offv", "オフボ"),
-        ("cover", "カバー"),
-        ("joke", "ネタ動画"),
-        ("other", "その他"),
-    )
-
-    title = models.CharField(default = "", max_length = 500)
-    authors = models.ManyToManyField('Author', related_name='songs', blank=True)
-    lyrics = models.TextField(blank = True, null = True, default = "", max_length = 10000)
-    imitates = models.ManyToManyField("self", symmetrical=False, related_name="imitateds", blank=True)
-    post_time = models.DateTimeField(default = timezone.now)
-    upload_time = models.DateTimeField(blank = True, null = True)
-    is_original = models.BooleanField(default = False)
-    is_joke = models.BooleanField(default = False)
-    is_deleted = models.BooleanField(default = False)
-    is_draft = models.BooleanField(default = False)
-    is_inst = models.BooleanField(default = False)
-    is_subeana = models.BooleanField(default = True)
-    is_special = models.BooleanField(default = False)
-    is_lock = models.BooleanField(default = False)
-    is_limited = models.BooleanField(default = False)
-    view = models.IntegerField(blank = True, null = True)
-    like = models.IntegerField(blank = True, null = True)
-    category = models.CharField(default = "song", choices=CHOICES, max_length=10)
-
-    def __str__(self):
-        return self.title
-
-    def authors_str(self, separator=", "):
-        """作者をカンマ区切りの文字列で返す"""
-        author_names = [author.name for author in self.authors.all()]
-        return separator.join(author_names)
-
-    def save(self, *args, **kwargs):
-        if self.lyrics:
-            self.lyrics = self.lyrics.replace("\r\n", "\n")
-        super().save(*args, **kwargs)
-
-
-# 曲のURLの情報
-class SongLink(models.Model):
-    url = models.URLField(max_length=500, unique=True)
-    songs = models.ManyToManyField('Song', blank=True, related_name='links')
-    is_removed = models.BooleanField(default=False)
-    allow_dup = models.BooleanField(default=False)
-
-    def __str__(self):
-        return self.url
-
-
-# 曲の作者の情報
-class Author(models.Model):
-    name = models.CharField(unique=True, max_length = 500)
-
-    def __str__(self):
-        return self.name
-
-
-# 曲の作者のwebページの情報
-class AuthorLink(models.Model):
-    url = models.CharField(max_length = 100)
-    author = models.ForeignKey(Author, on_delete = models.CASCADE, null=True, related_name="links")
-
-
-# 曲の作者の別の呼び方の情報
-# 曲の登録時や編集時に正式な呼び方(author.name)に変更するために使用される
-class AuthorAlias(models.Model):
-    CHOICES = (
-        ("id", "ID"),
-        ("abbr", "略称"),
-        ("common", "通称"),
-        ("past", "以前の名称"),
-        ("sns", "SNSでの名称"),
-        ("spell", "表記揺れ"),
-        ("other", "別名義"),
-    )
-
-    name = models.CharField(unique=True, max_length = 500)
-    alias_type = models.CharField(default = "other", choices=CHOICES, max_length=10)
-    author = models.ForeignKey(Author, on_delete = models.CASCADE, related_name="aliases")
-
-    def __str__(self):
-        return self.name
-
-
-# ユーザーによる曲や作者の変更を記録した情報
-# changeは編集内容をマークダウンで記録する
-class History(models.Model):
-    CHOICES = (
-        ("new", "新規作成"),
-        ("edit", "編集"),
-        ("delete", "削除申請"),
-    )
-
-    song = models.ForeignKey(Song, blank = True, null = True, on_delete = models.SET_NULL, related_name="histories")
-    title = models.CharField(default = "", max_length = 100)
-    history_type = models.CharField(default = "new", choices=CHOICES, max_length = 10)
-    create_time = models.DateTimeField(default = timezone.now)
-    changes = models.JSONField(null=True, blank=True, default=None)
-    editor = models.ForeignKey(Editor, on_delete = models.CASCADE, related_name="histories")
-
-
-# すべかしへの問い合わせとその回答の情報
-class Contact(models.Model):
-    detail = models.TextField(max_length = 10000)
-    post_time = models.DateField()
-    answer = models.TextField(blank = True, null = True, max_length = 10000)
-
-    def __str__(self):
-        return self.detail[:30]
-
-
-# 正常に狂うのです。が生成した歌詞の情報
-class Ai(models.Model):
-    lyrics = models.CharField(default = "", max_length = 100)
-    score = models.IntegerField(default = 0)
-    genetype = models.CharField(default = "", max_length = 100)
-
-    def __str__(self):
-        return self.lyrics
-
-
-# すべかしDSPに登録された曲の情報
-class Ad(models.Model):
-    CHOICES = (
-        ("still", "未審査"),
-        ("pass", "公開中"),
-        ("fail", "未通過"),
-    )
-    url = models.CharField(default = "", max_length = 100)
-    view = models.IntegerField(default = 0)
-    click = models.IntegerField(default = 0)
-    dup = models.IntegerField(default = 0)
-    status = models.CharField(default = "still", choices=CHOICES, max_length=10)
-
-    def __str__(self):
-        return self.url
