+++ subekashi/models/__init__.py
@@ -1,0 +1,8 @@
+from .editor import Editor
+from .author import Author, AuthorLink, AuthorAlias
+from .song import Song
+from .song_link import SongLink
+from .history import History
+from .contact import Contact
+from .ai import Ai
+from .ad import Ad
